% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colordiff.R
\name{cie76}
\alias{cie76}
\title{Get the CIE76 color difference between two CIELab values.}
\usage{
cie76(lab_color1, lab_color2)

cie76(lab_color1, lab_color2)
}
\arguments{
\item{lab_color1}{A vector with three components corresponding to a Lab value.}

\item{lab_color2}{A vector with three components corresponding to another Lab value.}

\item{color}{A matrix whose rows specify color coordinates in the CIELab color space.}

\item{reference}{A reference color.}

\item{...}{Weighting factors `k_L`, `k_C`, `k_H`, `K1`, and/or `K2` for CIE94 and CIEDE2000, if applicable. Also, `symmetric=TRUE` to use a symmetric version of CIE94.}
}
\value{
The CIE76 color difference between the two given values.

The CIEDE2000 color difference between the two given values.

The CIE76 color difference between the two given values.
}
\description{
Get the CIE76 color difference between two CIELab values.

Get the color difference between values in the CIELab color space.
}
\examples{

Get the CIE76 color difference between two CIELab values.

}
\references{
Sharma, G., & Bala, R. (Eds.). (2017). Digital color imaging handbook. CRC press.

Sharma, G., & Bala, R. (Eds.). (2017). Digital color imaging handbook. CRC press.
Sharma, G., Wu, W., & Dalal, E. N. (2005). The CIEDE2000 color‐difference formula: Implementation notes, supplementary test data, and mathematical observations. Color Research & Application: Endorsed by Inter‐Society Color Council, The Colour Group (Great Britain), Canadian Society for Color, Color Science Association of Japan, Dutch Society for the Study of Color, The Swedish Colour Centre Foundation, Colour Society of Australia, Centre Français de la Couleur, 30(1), 21-30.

Sharma, G., & Bala, R. (Eds.). (2017). Digital color imaging handbook. CRC press.
}
