% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name.R
\name{get_closest_color}
\alias{get_closest_color}
\title{Get information about the closest RHS color to some CIELab coordinates.}
\usage{
get_closest_color(L, a, b, metric = "CIEDE2000")
}
\arguments{
\item{L}{The lightness L* of the color.}

\item{a}{The chromatic component a* (red-green).}

\item{b}{The chromatic component b* (blue-yellow).}

\item{metric}{The color distance to use.}
}
\value{
A one-row tibble.
}
\description{
Get information about the closest RHS color to some CIELab coordinates.
}
\examples{
get_closest_color(65, 20, -20)
}
