% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/namer.R
\name{name_color}
\alias{name_color}
\title{Name a color given its coordinates in the CIELab color space.}
\usage{
name_color(L, a, b, language = "english")
}
\arguments{
\item{L}{The lightness L* of the color.}

\item{a}{The chromatic component a* (red-green).}

\item{b}{The chromatic component b* (blue-yellow).}

\item{language}{The language of the color name, between English, French, German, and Spanish.}
}
\value{
The name of the color, according to the UPOV.
}
\description{
Name a color given its coordinates in the CIELab color space.
}
\examples{
name_color(65, 20, -20)
name_color(65, 20, -20, "spanish")
name_color(65, 20, -20, "es")
}
