% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rhs_color_names_2015}
\alias{rhs_color_names_2015}
\title{UPOV names and groups for RHS colors.}
\format{
A data frame with 920 rows and 10 variables:
\describe{
\item{UPOVGroup}{the UPOV group of the color}
\item{RHS}{the RHS code of the color}
\item{english}{the English name for the color}
\item{french}{the French name for the color}
\item{german}{the German name for the color}
\item{spanish}{the Spanish name for the color}
}
}
\source{
UPOV
\url{https://www.upov.int/meetings/en/doc_details.jsp?meeting_id=50790&doc_id=426293}
}
\usage{
rhs_color_names_2015
}
\description{
Data set containing English, French, German, and Spanish names for the colors
defined by the RHS in its sixth edition (2015), alongside their UPOV group
number.
}
\keyword{datasets}
