% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rhs_color_values_2007}
\alias{rhs_color_values_2007}
\title{RHS colors in different color spaces.}
\format{
A data frame with 892 rows and 10 variables:
\describe{
\item{RHS}{the RHS code of the color}
\item{R}{the red component in sRGB}
\item{G}{the green component in sRGB}
\item{B}{the blue component in sRGB}
\item{L}{the lightness component in CIELab (D65 / 10º)}
\item{a}{the red-green component in CIELab (D65 / 10º)}
\item{b}{the blue-yellow component in CIELab (D65 / 10º)}
\item{L2}{the lightness component in CIELCh (D65 / 10º)}
\item{C}{the colorfulness component in CIELCh (D65 / 10º)}
\item{h}{the hue in CIELCh (D65 / 10º)}
}
}
\source{
\url{http://rhscf.orgfree.com/}
}
\usage{
rhs_color_values_2007
}
\description{
Data set containing the coordinates in RGB, CIELab, and CIELCh of the colors
defined by the Royal Horticultural Society in its fifth edition (2007).
}
\keyword{datasets}
