% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{upov}
\alias{upov}
\title{UPOV names and groups for RHS colors.}
\format{
A data frame with 920 rows and 10 variables:
\describe{
  \item{UPOVGroup}{the UPOV group of the color}
  \item{RHS}{the RHS code of the color}
  \item{english}{the English name for the color}
  \item{français}{the French name for the color}
  \item{deutsch}{the German name for the color}
  \item{español}{the Spanish name for the color}
}
}
\source{
\url{'https://www.upov.int/edocs/mdocs/upov/en/two_51/twp_3_11.docx}
}
\usage{
upov
}
\description{
Dataset containing English, French, German, and Spanish names for the colors
defined by the RHS in its sixth edition (2015), alongside their UPOV group
number.
}
\keyword{datasets}
