% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colordiff.R
\name{cie76}
\alias{cie76}
\title{Get the CIE76 color difference between two CIELab values.}
\usage{
cie76(lab_color1, lab_color2)
}
\arguments{
\item{lab_color1}{A vector with three components corresponding to a Lab value.}

\item{lab_color2}{A vector with three components corresponding to another Lab value.}
}
\value{
The CIE76 color difference between the two given values.
}
\description{
Get the CIE76 color difference between two CIELab values.
}
\references{
Sharma, G., & Bala, R. (Eds.). (2017). Digital color imaging handbook. CRC press.
}
